-display photos for more than 67 years in the history chart
-fade out photo tooltip immediatedly (without delay) when alpha is not fully opaque

+++ net.tourbook.device.garmin/src/net/tourbook/device/garmin/GarminSAXHandler.java
@@ -559,7 +559,7 @@ public class GarminSAXHandler extends DefaultHandler {
 			 * export and import found a wrong tour when exporting was done with camouflage speed ->
 			 * this will result in a NEW tour
 			 */
-			final int tourRecordingTime = tourData.getTourRecordingTime();
+			final int tourRecordingTime = (int) tourData.getTourRecordingTime();
 
 			if (distanceSerie == null) {
 				uniqueKey = Integer.toString(tourRecordingTime);
