+++ net.tourbook.device/src/net/tourbook/device/gpx/GPX_SAX_Handler.java
@@ -592,7 +592,7 @@ public class GPX_SAX_Handler extends DefaultHandler {
 			 * export and import found a wrong tour when exporting was done with camouflage speed ->
 			 * this will result in a NEW tour
 			 */
-			final int tourRecordingTime = tourData.getTourRecordingTime();
+			final int tourRecordingTime = (int) tourData.getTourRecordingTime();
 
 			if (distanceSerie == null) {
 				uniqueKey = Integer.toString(tourRecordingTime);
